fix: Add support for preserving mtime for replication
This PR is needed for bucket replication support
