+++ cmd/object-api-interface.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"io"
 	"net/http"
+	"time"
 
 	"github.com/minio/minio-go/v6/pkg/encrypt"
 	"github.com/minio/minio-go/v6/pkg/tags"
@@ -36,11 +37,12 @@ type GetObjectInfoFn func(ctx context.Context, bucket, object string, opts Objec
 // ObjectOptions represents object options for ObjectLayer object operations
 type ObjectOptions struct {
 	ServerSideEncryption encrypt.ServerSide
-	Versioned            bool
-	VersionID            string
-	UserDefined          map[string]string
-	PartNumber           int
-	CheckCopyPrecondFn   CheckCopyPreconditionFn
+	Versioned            bool                    // indicates if the bucket is versioned
+	VersionID            string                  // Specifies the versionID which needs to be overwritten or read
+	MTime                time.Time               // Is only set in POST/PUT operations
+	UserDefined          map[string]string       // only set in case of POST/PUT operations
+	PartNumber           int                     // only useful in case of GetObject/HeadObject
+	CheckCopyPrecondFn   CheckCopyPreconditionFn // only set during CopyObject preconditional valuation
 }
 
 // BucketOptions represents bucket options for ObjectLayer bucket operations
